refactor `data` module to be within modules, as it should be
