+++ docs/api/models/index.md
@@ -18,6 +18,7 @@ keep what is common common, and what is unique unique.
 ```{toctree}
 buffer
 config
+data
 mixins
 models
 sdcard
