+++ tests/fixtures.py
@@ -4,8 +4,7 @@ import pytest
 
 from miniscope_io import SDCard
 from miniscope_io.formats import WireFreeSDLayout
-from miniscope_io.data import Frames
-
+from miniscope_io.models.data import Frames
 
 
 @pytest.fixture
@@ -14,9 +13,10 @@ def wirefree() -> SDCard:
     SDCard with wirefree layout pointing to the sample data file
 
     """
-    sd_path = Path(__file__).parent.parent / 'data' / 'wirefree_example.img'
-    sdcard = SDCard(drive = sd_path, layout = WireFreeSDLayout)
+    sd_path = Path(__file__).parent.parent / "data" / "wirefree_example.img"
+    sdcard = SDCard(drive=sd_path, layout=WireFreeSDLayout)
     return sdcard
+
 
 @pytest.fixture()
 def wirefree_frames(wirefree) -> Frames:
